+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/DisabledVAD.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine.Scripting;
+
+namespace OpenAI.Realtime
+{
+    [Preserve]
+    public sealed class DisabledVAD : IVoiceActivityDetectionSettings
+    {
+        [Preserve]
+        public TurnDetectionType Type => TurnDetectionType.Disabled;
+
+        public bool CreateResponse => false;
+
+        public bool InterruptResponse => false;
+    }
+}
