+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/DisabledVAD.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3d1179c6e8165042863c648d853da68
+guid: abf9294316279b44bb6bbeb1a524a3b4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
