+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/IVoiceActivityDetectionSettings.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine.Scripting;
+
+namespace OpenAI.Realtime
+{
+    [Preserve]
+    public interface IVoiceActivityDetectionSettings
+    {
+        public TurnDetectionType Type { get; }
+        public bool CreateResponse { get; }
+        public bool InterruptResponse { get; }
+    }
+}
