+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/IVoiceActivityDetectionSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3d1179c6e8165042863c648d853da68
+guid: aaca36ccb4f1be849b16585f1163e4fb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
