+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/SemanticVAD.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3d1179c6e8165042863c648d853da68
+guid: b75fb647eaa84b246b1d13ce3e8f0627
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
