+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/ServerVAD.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3d1179c6e8165042863c648d853da68
+guid: 768baa7bd761c1e45ab7efd8176eb331
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
