+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/VAD_Eagerness.cs
@@ -1,0 +1,18 @@
+﻿using System.Runtime.Serialization;
+using UnityEngine.Scripting;
+
+namespace OpenAI.Realtime
+{
+    [Preserve]
+    public enum VAD_Eagerness
+    {
+        [EnumMember(Value = "auto")]
+        Auto = 0,
+        [EnumMember(Value = "low")]
+        Low,
+        [EnumMember(Value = "medium")]
+        Medium,
+        [EnumMember(Value = "high")]
+        High
+    }
+}
