+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/VAD_Eagerness.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3d1179c6e8165042863c648d853da68
+guid: cef010b5ac522234784b71cc178d8805
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
