+++ OpenAI/Packages/com.openai.unity/Runtime/Realtime/VoiceActivityDetectionSettings.cs
@@ -2,108 +2,10 @@
 
 using Newtonsoft.Json;
 using System;
-using System.Runtime.Serialization;
 using UnityEngine.Scripting;
 
 namespace OpenAI.Realtime
 {
-    [Preserve]
-    public interface IVoiceActivityDetectionSettings
-    {
-        public TurnDetectionType Type { get; }
-        public bool CreateResponse { get; }
-        public bool InterruptResponse { get; }
-    }
-
-    [Preserve]
-    public sealed class DisabledVAD : IVoiceActivityDetectionSettings
-    {
-        [Preserve]
-        public TurnDetectionType Type => TurnDetectionType.Disabled;
-
-        public bool CreateResponse => false;
-
-        public bool InterruptResponse => false;
-    }
-
-    [Preserve]
-    public enum VAD_Eagerness
-    {
-        [EnumMember(Value = "auto")]
-        Auto = 0,
-        [EnumMember(Value = "low")]
-        Low,
-        [EnumMember(Value = "medium")]
-        Medium,
-        [EnumMember(Value = "high")]
-        High
-    }
-
-    [Preserve]
-    public sealed class SemanticVAD : IVoiceActivityDetectionSettings
-    {
-        public SemanticVAD(bool createResponse = true, bool interruptResponse = true, VAD_Eagerness eagerness = VAD_Eagerness.Auto)
-        {
-            CreateResponse = createResponse;
-            InterruptResponse = interruptResponse;
-            Eagerness = eagerness;
-        }
-
-        [Preserve]
-        [JsonProperty("type", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public TurnDetectionType Type { get; } = TurnDetectionType.Semantic_VAD;
-
-        [JsonProperty("create_response", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public bool CreateResponse { get; private set; }
-
-        [JsonProperty("interrupt_response", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public bool InterruptResponse { get; private set; }
-
-        [Preserve]
-        [JsonProperty("eagerness", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public VAD_Eagerness Eagerness { get; private set; }
-    }
-
-    [Preserve]
-    public sealed class ServerVAD : IVoiceActivityDetectionSettings
-    {
-        public ServerVAD(
-            bool createResponse = true,
-            bool interruptResponse = true,
-            int? prefixPadding = null,
-            int? silenceDuration = null,
-            float? detectionThreshold = null)
-        {
-            CreateResponse = createResponse;
-            InterruptResponse = interruptResponse;
-            PrefixPadding = prefixPadding;
-            SilenceDuration = silenceDuration;
-            DetectionThreshold = detectionThreshold;
-        }
-
-        [Preserve]
-        [JsonProperty("type", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public TurnDetectionType Type { get; } = TurnDetectionType.Server_VAD;
-
-        [JsonProperty("create_response", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public bool CreateResponse { get; private set; }
-
-        [JsonProperty("interrupt_response", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public bool InterruptResponse { get; private set; }
-
-        [Preserve]
-        [JsonProperty("prefix_padding_ms", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int? PrefixPadding { get; private set; }
-
-        [Preserve]
-        [JsonProperty("silence_duration_ms", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public int? SilenceDuration { get; private set; }
-
-        [Preserve]
-        [JsonProperty("threshold", DefaultValueHandling = DefaultValueHandling.Ignore)]
-        public float? DetectionThreshold { get; private set; }
-    }
-
     [Preserve]
     [Obsolete("Use new IVoiceActivityDetectionSettings classes: SemanticVAD, ServerVAD, and DisabledVAD")]
     public sealed class VoiceActivityDetectionSettings : IVoiceActivityDetectionSettings
